feat: replace data ferching into separate function
